Feature/oe 5628

+++ protected/behaviors/ContactBehavior.php
@@ -169,7 +169,7 @@ class ContactBehavior extends CActiveRecordBehavior
 
 	public function afterDelete($event)
 	{
-		if ($this->owner->contact_id) {
+		if ($this->owner->contact_id && get_class($this->owner) != 'ContactLocation') {
 			Address::model()->deleteAll('contact_id = :contact_id', array(':contact_id' => $this->owner->contact_id));
 			Contact::model()->deleteByPk($this->owner->contact_id);
 		}
